Uploading an old project I made in my free time while at COSMOS in 2017.
The project allowed me to move my mouse with my violin, with 2 strings
controlling the x-axis and 2 strings controllin the y-axis.
Ultimately, the project got scrapped. My main project with partners had
started. My partners had no programming experience, so I had to work on the glvoe mostly by myself. I also had signficiant problems with socket connections being unstable, which was too much for my 14-year-old brain to work around.
I would try to fix it and get it to work, but my knowledge of PD, which
is a super niche language to begin with (I only used it because it was
being taught at the program at the time) has completely disappeared.
Maybe if I ever use PD in the future, I will go back and fix this
project. For now, I just want to get it onto Git.